Move events services in new ui services directory
