+++ ui/main/src/app/business/application-loader.ts
@@ -19,7 +19,7 @@ import {EntitiesService} from '../services/entities/EntitiesService';
 import {GroupsService} from '../services/groups/GroupsService';
 import {PerimetersService} from '../services/perimeters/PerimetersService';
 import {ProcessesService} from '../services/processes/ProcessesService';
-import {OpfabEventStreamService} from './services/events/opfabEventStream.service';
+import {OpfabEventStreamService} from '../services/events/OpfabEventStreamService';
 import {AcknowledgeService} from '../services/acknowlegment/AcknowledgeService';
 import {DateTimeFormatterService} from '../services/dateTimeFormatter/DateTimeFormatterService';
 import {AdminProcessesService} from '../services/admin/AdminProcessesService';
@@ -33,7 +33,7 @@ import {TemplateCssService} from '../services/templateCss/TemplateCssService';
 import {UserSettingsService} from '../services/userSettings/UserSettingsService';
 import {NavigationService} from '../services/navigation/NavigationService';
 import {OpfabStore} from './store/opfabStore';
-import {ApplicationUpdateService} from './services/events/application-update.service';
+import {ApplicationUpdateService} from '../services/events/ApplicationUpdateService';
 import {SystemNotificationService} from '../services/notifications/SystemNotificationService';
 import {ApplicationLoadingComponent} from './application-loading-component';
 import {ServerResponseStatus} from './server/serverResponse';
