+++ ui/main/src/app/business/store/lightcards/lightcards-store.ts
@@ -23,11 +23,11 @@ import {
     tap
 } from 'rxjs';
 import {UsersService} from '@ofServices/users/UsersService';
-import {OpfabEventStreamService} from 'app/business/services/events/opfabEventStream.service';
-import {CardOperationType} from '@ofModel/card-operation.model';
+import {OpfabEventStreamService} from '@ofServices/events/OpfabEventStreamService';
 import {LogOption, LoggerService as logger} from 'app/services/logs/LoggerService';
 import {SelectedCardStore} from 'app/business/store/selectedCard.store';
 import {AcknowledgeService} from '../../../services/acknowlegment/AcknowledgeService';
+import {CardOperationType} from '@ofServices/events/model/CardOperation';
 
 /**
  *
