+++ ui/main/src/app/business/view/activityarea/activityarea.view.ts
@@ -1,4 +1,4 @@
-/* Copyright (c) 2023-2024, RTE (http://www.rte-france.com)
+/* Copyright (c) 2023-2025, RTE (http://www.rte-france.com)
  * See AUTHORS.txt
  * This Source Code Form is subject to the terms of the Mozilla Public
  * License, v. 2.0. If a copy of the MPL was not distributed with this
@@ -17,7 +17,7 @@ import {ActivityAreaEntityCluster, ActivityAreaLine, ActivityAreaPage} from './a
 import {RoleEnum} from '@ofServices/entities/model/RoleEnum';
 import {OpfabStore} from 'app/business/store/opfabStore';
 import {Entity} from '@ofServices/entities/model/Entity';
-import {ApplicationEventsService} from 'app/business/services/events/application-events.service';
+import {ApplicationEventsService} from '@ofServices/events/ApplicationEventsService';
 
 export class ActivityAreaView {
     private readonly activityAreaSubject = new ReplaySubject<ActivityAreaPage>(1);
