+++ ui/main/src/app/modules/card/components/card-footer-text/card-footer-text.component.ts
@@ -17,7 +17,7 @@ import {Utilities} from 'app/business/common/utilities';
 import {Subject} from 'rxjs';
 import {takeUntil} from 'rxjs/operators';
 import {OpfabStore} from 'app/business/store/opfabStore';
-import {CardOperationType} from '@ofModel/card-operation.model';
+import {CardOperationType} from '@ofServices/events/model/CardOperation';
 import {TranslateModule} from '@ngx-translate/core';
 import {NgIf, NgFor} from '@angular/common';
