+++ ui/main/src/tests/mocks/opfab-event-stream.server.mock.ts
@@ -1,4 +1,4 @@
-/* Copyright (c) 2023-2024, RTE (http://www.rte-france.com)
+/* Copyright (c) 2023-2025, RTE (http://www.rte-france.com)
  * See AUTHORS.txt
  * This Source Code Form is subject to the terms of the Mozilla Public
  * License, v. 2.0. If a copy of the MPL was not distributed with this
@@ -7,9 +7,9 @@
  * This file is part of the OperatorFabric project.
  */
 
-import {CardOperation} from '@ofModel/card-operation.model';
+import {CardOperation} from '@ofServices/events/model/CardOperation';
 import {LightCard} from '@ofModel/light-card.model';
-import {OpfabEventStreamServer} from 'app/business/server/opfabEventStream.server';
+import {OpfabEventStreamServer} from '@ofServices/events/server/OpfabEventStreamServer';
 import {ServerResponse, ServerResponseStatus} from 'app/business/server/serverResponse';
 import {Observable, of, Subject} from 'rxjs';
